+++ symfony/plugins/orangehrmAdminPlugin/modules/admin/actions/deletePayGradesAction.class.php
@@ -45,10 +45,15 @@ class deletePayGradesAction extends baseAdminAction {
                     foreach ($toBeDeletedPayGradeIds as $toBeDeletedPayGradeId) {
 
                         $payGrade = $this->getPayGradeService()->getPayGradeById($toBeDeletedPayGradeId);
-                        $payGrade->delete();
+                        if ($payGrade instanceof PayGrade) {
+                            $payGrade->delete();
+                        }
                     }
                     $this->getUser()->setFlash('success', __(TopLevelMessages::DELETE_SUCCESS));
                 }
+            } else {
+                $this->handleBadRequest();
+                $this->forwardToSecureAction();
             }
 
             $this->redirect('admin/viewPayGrades');
